Addressing feedback on PR
