+++ examples/blueprint-construct/index.ts
@@ -177,10 +177,10 @@ export default class BlueprintConstruct {
               virtualClusterName: 'batchJob',
               virtualClusterNamespace: 'batchjob',
               createNamespace: true,
-              excutionRoles: [
+              executionRoles: [
                   {
-                      excutionRoleIamPolicyStatement: executionRolePolicyStatement,
-                      excutionRoleName: 'myBlueprintExecRole'
+                      executionRoleIamPolicyStatement: executionRolePolicyStatement,
+                      executionRoleName: 'myBlueprintExecRole'
                   }
               ]
           };
