--- lib/teams/emrContainersRoleBinding.yaml
@@ -1,13 +1,0 @@
-apiVersion: rbac.authorization.k8s.io/v1
-kind: RoleBinding
-metadata:
-  name: emr-containers
-  namespace: <REPLACE-NAMESPACE>
-subjects:
-- kind: User
-  name: emr-containers
-  apiGroup: rbac.authorization.k8s.io
-roleRef:
-  kind: Role
-  name: emr-containers
-  apiGroup: rbac.authorization.k8s.io
